requirements.txt
flask
openai
speechrecognition
pyttsx3
google-generativeai